Allow user to specify buffering mode of capture file
It can be useful to allow the user to avoid the overhead of a write(2)
for every character printed to the capture file. To that end, add a
new option --capturefile-buffer-mode=[NLF], which translate
directly to the _IONBF, _IOLBF, _IOFBF used with setvbuf(3).

For backward compatibility, the default is of course unbuffered aka
_IONBF.

I don't see much of a need to allow changing the buffering mode at
run-time, though it shouldn't be too hard to implement. But it would
only take effect after re-opening the capture file (since setvbuf must
only be called on a fresh FILE*), so perhaps one should only be
allowed to change capbuf when capfp is NULL.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>