Change mc_wdrawelm_var to allocate buffer
Allocating the buffer outside of the function
does not allow to alloc the right size and it
will eventually be to small. Also it was
not told to mc_wdrawelm_var how large the buffer
is. Callers of mc_wdrawelm_var need to free
the buffer again after use now.