Rework device file locking
Add a third locking mode that kicks in when the file-based locking does
not work, e.g., because the lock directory is not writable. The new
third method uses advisory locks. Other programs using different kind of
locks will not check for the lock and use it anyway. That's why we keep
using the file-based one.

This addresses #9.